Hotel data + Filter by price, reviews, city

+++ Web Api/Users/Users/Entities/Dto/Hotels/ReviewForCreationDto.cs
@@ -6,7 +6,7 @@ namespace Users.Entities.Dto.Hotels
         public int Id { get; set; }
         public int UserId { get; set; }
         public int HotelId { get; set; }
-        public byte ReviewStar { get; set; }
+        public decimal ReviewStar { get; set; }
         public string? Description { get; set; }
         public DateTime CreatedDate { get; set; } = DateTime.Now;
     }
