+++ Web Api/Users/Users/Entities/Dto/Hotels/ReviewForUpdateDro.cs
@@ -4,7 +4,7 @@ namespace Users.Entities.Dto.Hotels
     public class ReviewForUpdateDto
     {
         public int Id { get; set; }
-        public byte ReviewStar { get; set; }
+        public decimal ReviewStar { get; set; }
         public string? Description { get; set; }
     }
 }
