+++ Web Api/Users/Users/Entities/Models/Review.cs
@@ -6,7 +6,7 @@ namespace Users.Entities.Models
         public int Id { get; set; }
         public int UserId { get; set; }
         public int HotelId { get; set; }
-        public byte ReviewStar { get; set; }
+        public decimal ReviewStar { get; set; }
         public string? Description { get; set; }
         public DateTime CreatedDate { get; set; }
         // public int RoomId { get; set; }
